requirements.txt
tensorflow==2.18.0
numpy==2.0.2
psycopg2-binary==2.9.10
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django
docopt==0.6.2
idna==3.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
yarg==0.1.9
channels==3.0.5
django-crispy-forms
Pillow
crispy-bootstrap4
